requirements.txt
black==24.8.0
boto3==1.35.31
botocore==1.35.31
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
idna==3.10
jmespath==1.0.1
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
s3transfer==0.10.2
six==1.16.0
urllib3==2.2.3
